WIP: draft of completing proper TxOut datum support
The datums are now annotated into the transaction's TxOut values.

This change is a bit tricky since TxOut is used both in the TxBody
(where we want to annotate the datums) and also in the UTxO. We resolve
this with an extra context type arg to TxOut.

TODO: split this commit into two: the API changes and the CLI changes.

TODO: rebase on top of the other API & CLI changes for automation of tx
building.